pixpl/requirements.txt
anyio==4.10.0
asgiref==3.9.1
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
futures==3.0.5
goslate==1.5.4
gunicorn==23.0.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpx==0.24.1
httpcore>=0.15.0,<0.18.0
h11>=0.13.0,<0.15.0
hyperframe==5.2.0
inflection==0.5.1
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
PyMySQL==1.1.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rfc3986==1.5.0
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
